Refactor to have one view per file
