+++ changelog.d/refactor-moving-partner-views-common-constructor-to-super-class.feature
@@ -1,0 +1,1 @@
+Refactor moving partner views constructor to super class
