+++ changelog.d/refactor-one-view-per-file.feature
@@ -1,0 +1,1 @@
+Refactor moving one view class per file
